// src/apis/models/RiskLevelRuleParameters.ts
/**
 * Tarpon Console API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

export class RiskLevelRuleParameters {
  'VERY_HIGH': any;
  'HIGH': any;
  'MEDIUM': any;
  'LOW': any;
  'VERY_LOW': any;

  static readonly discriminator: string | undefined = undefined;

  static readonly attributeTypeMap: Array<{
    name: string;
    baseName: string;
    type: string;
    format: string;
  }> = [
    {
      name: 'VERY_HIGH',
      baseName: 'VERY_HIGH',
      type: 'any',
      format: '',
    },
    {
      name: 'HIGH',
      baseName: 'HIGH',
      type: 'any',
      format: '',
    },
    {
      name: 'MEDIUM',
      baseName: 'MEDIUM',
      type: 'any',
      format: '',
    },
    {
      name: 'LOW',
      baseName: 'LOW',
      type: 'any',
      format: '',
    },
    {
      name: 'VERY_LOW',
      baseName: 'VERY_LOW',
      type: 'any',
      format: '',
    },
  ];

  static getAttributeTypeMap() {
    return RiskLevelRuleParameters.attributeTypeMap;
  }

  public constructor() {}
}